requirements_full.txt
altair==4.1.0
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.16.0
botocore==1.19.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.9
heroku==0.1.4
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
opencv-python==4.4.0.44
opencv-python-headless==4.4.0.44
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pdf2image==1.14.0
pdfminer.six==20200517
pdfplumber==0.5.24
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
pyarrow==2.0.0
pycparser==2.20
pycryptodome==3.9.9
pydeck==0.5.0b1
Pygments==2.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
sortedcontainers==2.2.2
streamlit==0.69.2
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tox==3.20.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.25.11
validators==0.18.1
virtualenv==20.0.35
Wand==0.6.3
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
